pass image paths as props in navigation
