+++ projects/Mallard/src/components/article/types/article.tsx
@@ -18,6 +18,10 @@ import {
 } from '../../../../../Apps/common/src'
 import { navigateToLightbox } from 'src/navigation/helpers/base'
 import { fetchLightboxSetting } from 'src/helpers/settings/debug'
+import { selectImagePath } from 'src/hooks/use-image-paths'
+import { useApiUrl } from 'src/hooks/use-settings'
+import { useIssueSummary } from 'src/hooks/use-issue-summary'
+import { Image } from 'src/common'
 
 const styles = StyleSheet.create({
     block: {
@@ -133,6 +137,7 @@ const Article = ({
     const [, { type }] = useArticle()
     const ref = useRef<WebView | null>(null)
     const [lightboxEnabled, setLightboxEnabled] = useState(false)
+    const [imagePaths, setImagePaths] = useState([''])
 
     const wasShowingHeader = useUpdateWebviewVariable(
         ref,
@@ -141,12 +146,42 @@ const Article = ({
     )
 
     const [, { pillar }] = useArticle()
+    const apiUrl = useApiUrl() || ''
+    const { issueId } = useIssueSummary()
 
     useEffect(() => {
         fetchLightboxSetting().then(lightboxEnabled =>
             setLightboxEnabled(lightboxEnabled),
         )
     }, [])
+
+    useEffect(() => {
+        const lbimages = getLightboxImages(article.elements)
+        const lbCreditedImages = getCreditedImages(lbimages)
+        if (article.type !== 'gallery' && article.image) {
+            lbCreditedImages.unshift(article.image)
+        }
+        const getImagePathFromImage = async (image: Image) => {
+            if (issueId && image) {
+                const { localIssueId, publishedIssueId } = issueId
+                const imagePath = await selectImagePath(
+                    apiUrl,
+                    localIssueId,
+                    publishedIssueId,
+                    image,
+                    'full-size',
+                )
+                return imagePath
+            }
+            return ''
+        }
+        const fetchImagePaths = async () => {
+            return await Promise.all(
+                lbCreditedImages.map(image => getImagePathFromImage(image)),
+            )
+        }
+        fetchImagePaths().then(imagePaths => setImagePaths(imagePaths))
+    }, [apiUrl, article.elements, issueId, article.image, article.type])
 
     return (
         <Fader>
@@ -208,6 +243,7 @@ const Article = ({
                             navigation,
                             navigationProps: {
                                 images: lbCreditedImages,
+                                imagePaths: imagePaths,
                                 index,
                                 pillar,
                             },
