+++ projects/Mallard/src/navigation/helpers/base.tsx
@@ -126,6 +126,7 @@ const navigateToSettings = (navigation: NavigationScreenProp<{}>): void => {
 
 export interface LightboxNavigationProps {
     images?: CreditedImage[]
+    imagePaths?: string[]
     index?: number
     pillar?: ArticlePillar
 }
