+++ projects/Mallard/src/screens/article/slider/index.tsx
@@ -89,7 +89,7 @@ const ArticleSlider = React.memo(
                     index: current,
                     animated: false,
                 })
-        }, [width]) // eslint-disable-line react-hooks/exhaustive-deps
+        }, [width, navigation.isFocused()]) // eslint-disable-line react-hooks/exhaustive-deps
 
         const { panResponder } = useDismissArticle()
